feat: remove the creation of cronjobs in cleanup controller (#8526)
* feat: remove the creation of cronjobs in cleanup controller


* fix: use lastExecutionTime instead of nextExecutionTime


---------
