+++ api/kyverno/v2alpha1/cleanup_policy_types.go
@@ -184,7 +184,8 @@ type CleanupPolicySpec struct {
 
 // CleanupPolicyStatus stores the status of the policy.
 type CleanupPolicyStatus struct {
-	Conditions []metav1.Condition `json:"conditions,omitempty" patchStrategy:"merge" patchMergeKey:"type" protobuf:"bytes,1,rep,name=conditions"`
+	Conditions        []metav1.Condition `json:"conditions,omitempty" patchStrategy:"merge" patchMergeKey:"type" protobuf:"bytes,1,rep,name=conditions"`
+	LastExecutionTime metav1.Time        `json:"lastExecutionTime,omitempty"`
 }
 
 // Validate implements programmatic validation
