+++ config/crds/kyverno.io_cleanuppolicies.yaml
@@ -1248,6 +1248,9 @@ spec:
                   - type
                   type: object
                 type: array
+              lastExecutionTime:
+                format: date-time
+                type: string
             type: object
         required:
         - spec
